Use size_t when converting integer from/to pointer
Perviously we use uint32_t for this, but this doesn't compile for 64-bit
environment (and likely doesn't for 16-bit either.) Use size_t so that
we don't get size mismatch errors.

BUG=chrome-os-partner:19257
TEST=Run host emulated tests
BRANCH=None

+++ common/util.c
@@ -196,8 +196,7 @@ void *memset(void *dest, int c, int len)
 
 void *memmove(void *dest, const void *src, int len)
 {
-	if ((uint32_t)dest <= (uint32_t)src ||
-	    (uint32_t)dest >= (uint32_t)src + len) {
+	if ((size_t)dest <= (size_t)src || (size_t)dest >= (size_t)src + len) {
 		/* Start of destination doesn't overlap source, so just use
 		 * memcpy(). */
 		return memcpy(dest, src, len);
